Increase test coverage.  Migrate some tests to new test_module.py
